addcommand_fixbugs

+++ WpfAppForGit/Commons/Currency.cs
@@ -10,7 +10,7 @@ namespace WpfAppForGit.Commons
     {
         public int r030 { get; set; }
         public string txt { get; set; }
-        public double rate { get; set; }
+        public decimal rate { get; set; }
         public string cc { get; set; }
         public string exchangedate { get; set; }
     }
